requirements_project.txt
# requirements_project.txt - Only additional packages needed
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.3.0
opencv-python==4.8.1.78
PyYAML==6.0.1
jupyter==1.0.0
tensorflow-hub==0.13.0
plotly==5.17.0